chore: extract DB fixtures into src/fixtures.ts
Part of REAL-1037.
