+++ test/e2e/realtime-check.ts
@@ -4,7 +4,7 @@ import { createClient, SupabaseClient, postgresChangesFilter } from "@supabase/s
 import kleur from "kleur";
 import { SQL } from "bun";
 import {
-  ANON_KEY, SERVICE_KEY, dbPassword, EMAIL_DOMAIN, JSON_OUTPUT, DB_URL_ARG, env,
+  ANON_KEY, SERVICE_KEY, dbPassword, JSON_OUTPUT, DB_URL_ARG, env,
   TEST_CATEGORIES, PROJECT_URL, DB_URL, DB_SSL, REALTIME_OPTS, BROADCAST_CONFIG, EVENT_TIMEOUT_MS,
   RATE_LIMIT_PAUSE_MS, BROADCAST_API_HEADERS, LOAD_MESSAGES, LOAD_SETTLE_MS, LOAD_DELIVERY_SLO,
 } from "./src/context.ts";
@@ -16,198 +16,7 @@ import {
   openChannel, openPostgresChannel, REPLICATION_READY_CONFIG, openReplicationChannel,
   executeInsert, executeUpdate, executeDelete,
 } from "./src/helpers.ts";
-
-const fmtSqlResult = (result: any[]) => {
-  const count = (result as any).count ?? result.length;
-  return result.length > 0 ? `count=${count} rows=${JSON.stringify(result)}` : `count=${count}`;
-};
-const runSql = (label: string, query: Promise<any[]>): Promise<any[]> =>
-  query
-    .then((r) => { log(kleur.dim(`setup:   ${label} ok (${fmtSqlResult(r)})`)); return r; })
-    .catch((e: unknown) => { log(kleur.red(`setup:   ${label} FAILED: ${e instanceof Error ? e.message : String(e)}`)); throw e; });
-
-async function setup(): Promise<{ userId: string; testUser: { email: string; password: string }; supabase: SupabaseClient }> {
-  const start = performance.now();
-  const email = `realtime-check-${crypto.randomUUID()}@${EMAIL_DOMAIN}`;
-  const password = crypto.randomUUID();
-
-  log("setup: connecting to database");
-  const sql = new SQL(DB_URL, { tls: DB_SSL || undefined });
-  let userId: string;
-  try {
-    let stepStart = performance.now();
-    log(kleur.dim("setup: truncating existing tables"));
-    await Promise.allSettled([
-      sql`TRUNCATE TABLE public.pg_changes, public.dummy, public.authorization, public.broadcast_changes, public.replay_check`.then(
-        () => log(kleur.dim("setup:   truncate ok")),
-        (e: unknown) => log(kleur.dim(`setup:   truncate skipped (${e instanceof Error ? e.message : String(e)})`))
-      ),
-    ]);
-    log(kleur.dim(`setup: truncate done (${(performance.now() - stepStart).toFixed(0)}ms)`));
-
-    stepStart = performance.now();
-    log(kleur.dim("setup: creating tables"));
-    await Promise.allSettled([
-      runSql("table pg_changes", sql`CREATE TABLE IF NOT EXISTS public.pg_changes (
-            id bigint GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY,
-            value text NOT NULL DEFAULT gen_random_uuid(),
-            details text
-          )`),
-      runSql("table dummy", sql`CREATE TABLE IF NOT EXISTS public.dummy (
-            id bigint GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY,
-            value text NOT NULL DEFAULT gen_random_uuid()
-          )`),
-      runSql("table authorization", sql`CREATE TABLE IF NOT EXISTS public.authorization (
-            id bigint GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY,
-            value text NOT NULL DEFAULT gen_random_uuid()
-          )`),
-      runSql("table broadcast_changes", sql`CREATE TABLE IF NOT EXISTS public.broadcast_changes (id text PRIMARY KEY, value text NOT NULL, topic text NOT NULL)`),
-      runSql("table wallet", sql`CREATE TABLE IF NOT EXISTS public.wallet (id text PRIMARY KEY, wallet_id text NOT NULL)`),
-      runSql("table replay_check", sql`CREATE TABLE IF NOT EXISTS public.replay_check (
-            id text PRIMARY KEY,
-            topic text NOT NULL,
-            event text NOT NULL,
-            payload jsonb NOT NULL DEFAULT '{}',
-            binary_payload bytea
-          )`),
-    ]);
-    await runSql("pg_changes details column", sql`ALTER TABLE public.pg_changes ADD COLUMN IF NOT EXISTS details text`);
-    await runSql("pg_changes nullable_value column", sql`ALTER TABLE public.pg_changes ADD COLUMN IF NOT EXISTS nullable_value text`);
-    await runSql("pg_changes replica identity", sql`ALTER TABLE public.pg_changes REPLICA IDENTITY FULL`);
-    await runSql("replay_check binary_payload column", sql`ALTER TABLE public.replay_check ADD COLUMN IF NOT EXISTS binary_payload bytea`);
-    log(kleur.dim(`setup: tables done (${(performance.now() - stepStart).toFixed(0)}ms)`));
-
-    stepStart = performance.now();
-    log(kleur.dim("setup: configuring RLS and publications"));
-    await Promise.allSettled([
-      runSql("wallet seed", sql`INSERT INTO public.wallet (id, wallet_id) VALUES ('1', 'wallet_1') ON CONFLICT (id) DO NOTHING`),
-      runSql("dummy RLS disable", sql`ALTER TABLE public.dummy DISABLE ROW LEVEL SECURITY`),
-      runSql("pg_changes RLS enable", sql`ALTER TABLE public.pg_changes ENABLE ROW LEVEL SECURITY`),
-      runSql("authorization RLS enable", sql`ALTER TABLE public.authorization ENABLE ROW LEVEL SECURITY`),
-      runSql("broadcast_changes RLS enable", sql`ALTER TABLE public.broadcast_changes ENABLE ROW LEVEL SECURITY`),
-      runSql("wallet RLS enable", sql`ALTER TABLE public.wallet ENABLE ROW LEVEL SECURITY`),
-      runSql("replay_check RLS enable", sql`ALTER TABLE public.replay_check ENABLE ROW LEVEL SECURITY`),
-      sql`ALTER PUBLICATION supabase_realtime ADD TABLE public.pg_changes`
-        .then((r) => log(kleur.dim(`setup:   publication pg_changes ok (${fmtSqlResult(r)})`)))
-        .catch((e: unknown) => log(kleur.dim(`setup:   publication pg_changes skipped (${e instanceof Error ? e.message : String(e)})`))),
-      sql`ALTER PUBLICATION supabase_realtime ADD TABLE public.dummy`
-        .then((r) => log(kleur.dim(`setup:   publication dummy ok (${fmtSqlResult(r)})`)))
-        .catch((e: unknown) => log(kleur.dim(`setup:   publication dummy skipped (${e instanceof Error ? e.message : String(e)})`))),
-    ]);
-    log(kleur.dim(`setup: RLS and publications done (${(performance.now() - stepStart).toFixed(0)}ms)`));
-
-    stepStart = performance.now();
-    log(kleur.dim("setup: creating policies"));
-    await Promise.allSettled([
-      runSql("policy 'authenticated receive on topic'", sql`DO $$ BEGIN
-            IF NOT EXISTS (SELECT 1 FROM pg_policies WHERE policyname = 'authenticated receive on topic' AND tablename = 'messages' AND schemaname = 'realtime') THEN
-              CREATE POLICY "authenticated receive on topic" ON "realtime"."messages" AS PERMISSIVE
-                FOR SELECT TO authenticated USING (realtime.topic() like 'topic:%');
-            END IF;
-          END $$`),
-      runSql("policy 'authenticated broadcast on topic'", sql`DO $$ BEGIN
-            IF NOT EXISTS (SELECT 1 FROM pg_policies WHERE policyname = 'authenticated broadcast on topic' AND tablename = 'messages' AND schemaname = 'realtime') THEN
-              CREATE POLICY "authenticated broadcast on topic" ON "realtime"."messages" AS PERMISSIVE
-                FOR INSERT TO authenticated WITH CHECK (realtime.topic() like 'topic:%');
-            END IF;
-          END $$`),
-      runSql("policy 'allow authenticated users all access'", sql`DO $$ BEGIN
-            IF NOT EXISTS (SELECT 1 FROM pg_policies WHERE policyname = 'allow authenticated users all access' AND tablename = 'pg_changes' AND schemaname = 'public') THEN
-              CREATE POLICY "allow authenticated users all access" ON "public"."pg_changes" AS PERMISSIVE
-                FOR ALL TO authenticated USING (TRUE);
-            END IF;
-          END $$`),
-      runSql("policy 'authenticated have full access to read on broadcast_changes'", sql`DO $$ BEGIN
-            IF NOT EXISTS (SELECT 1 FROM pg_policies WHERE policyname = 'authenticated have full access to read on broadcast_changes' AND tablename = 'broadcast_changes' AND schemaname = 'public') THEN
-              CREATE POLICY "authenticated have full access to read on broadcast_changes" ON "public"."broadcast_changes" AS PERMISSIVE
-                FOR ALL TO authenticated USING (TRUE);
-            END IF;
-          END $$`),
-      runSql("policy 'authenticated have full access to replay_check'", sql`DO $$ BEGIN
-            IF NOT EXISTS (SELECT 1 FROM pg_policies WHERE policyname = 'authenticated have full access to replay_check' AND tablename = 'replay_check' AND schemaname = 'public') THEN
-              CREATE POLICY "authenticated have full access to replay_check" ON "public"."replay_check" AS PERMISSIVE
-                FOR ALL TO authenticated USING (TRUE) WITH CHECK (TRUE);
-            END IF;
-          END $$`),
-    ]);
-    log(kleur.dim(`setup: policies done (${(performance.now() - stepStart).toFixed(0)}ms)`));
-
-    stepStart = performance.now();
-    log(kleur.dim("setup: creating functions and triggers"));
-    await runSql("function broadcast_changes_for_table_trigger", sql`
-      CREATE OR REPLACE FUNCTION broadcast_changes_for_table_trigger() RETURNS TRIGGER AS $$
-      DECLARE topic text;
-      BEGIN
-        topic = COALESCE(NEW.topic, OLD.topic);
-        PERFORM realtime.broadcast_changes(topic, TG_OP, TG_OP, TG_TABLE_NAME, TG_TABLE_SCHEMA, NEW, OLD, TG_LEVEL);
-        RETURN NULL;
-      END;
-      $$ LANGUAGE plpgsql
-    `);
-    await runSql("broadcast_changes topic column", sql`ALTER TABLE public.broadcast_changes ADD COLUMN IF NOT EXISTS topic text NOT NULL`);
-
-    await runSql("trigger broadcast_changes_for_table_public_broadcast_changes_trigger", sql`
-      DO $$ BEGIN
-        IF NOT EXISTS (SELECT 1 FROM pg_trigger WHERE tgname = 'broadcast_changes_for_table_public_broadcast_changes_trigger') THEN
-          CREATE TRIGGER broadcast_changes_for_table_public_broadcast_changes_trigger
-            AFTER INSERT OR UPDATE OR DELETE ON broadcast_changes
-            FOR EACH ROW EXECUTE FUNCTION broadcast_changes_for_table_trigger();
-        END IF;
-      END $$
-    `);
-
-    await runSql("function replay_check_trigger", sql`
-      CREATE OR REPLACE FUNCTION replay_check_trigger() RETURNS TRIGGER AS $$
-      BEGIN
-        IF NEW.binary_payload IS NULL THEN
-          PERFORM realtime.send(NEW.payload, NEW.event, NEW.topic, true);
-        ELSE
-          PERFORM realtime.send_binary(NEW.binary_payload, NEW.event, NEW.topic, true);
-        END IF;
-        RETURN NULL;
-      END;
-      $$ LANGUAGE plpgsql
-    `);
-
-    await runSql("trigger replay_check_send_trigger", sql`
-      DO $$ BEGIN
-        IF NOT EXISTS (SELECT 1 FROM pg_trigger WHERE tgname = 'replay_check_send_trigger') THEN
-          CREATE TRIGGER replay_check_send_trigger
-            AFTER INSERT ON public.replay_check
-            FOR EACH ROW EXECUTE FUNCTION replay_check_trigger();
-        END IF;
-      END $$
-    `);
-
-    log(kleur.dim(`setup: functions and triggers done (${(performance.now() - stepStart).toFixed(0)}ms)`));
-
-    log(kleur.dim("setup: creating test user"));
-    const admin = createClient(PROJECT_URL, SERVICE_KEY);
-    const { data, error } = await admin.auth.admin.createUser({ email, password, email_confirm: true });
-    if (error) throw new Error(`Failed to create test user: ${error.message}`);
-    userId = data.user.id;
-    log(kleur.dim(`setup: done (${(performance.now() - start).toFixed(0)}ms)`));
-  } finally {
-    await sql.close().catch(() => {});
-  }
-
-  const supabase = createClient(PROJECT_URL, ANON_KEY, { realtime: REALTIME_OPTS });
-  await signInUser(supabase, email, password);
-  return { userId: userId!, testUser: { email, password }, supabase };
-}
-
-async function cleanup(userId: string) {
-  log("cleanup: deleting test user");
-  const sql = new SQL(DB_URL, { tls: DB_SSL || undefined });
-  try {
-    await sql`DELETE FROM auth.users WHERE id = ${userId}`;
-    log(kleur.dim("cleanup: done"));
-  } catch (_e) {
-    log(kleur.yellow("Warning: failed to clean up test user"));
-  } finally {
-    await sql.close().catch(() => {});
-  }
-}
+import { setup, cleanup } from "./src/fixtures.ts";
 
 async function runConnectionTest() {
   suite("connection");
